+++ NetworkKMM/docs/changelog.md
@@ -3,6 +3,7 @@
 ##### Unreleased
 
 - feature: support custom HTTP methods via VBTransportRequest
+- feature: add multipart/form-data request body builder
 - docs: add PUT and binary upload examples
 
 ##### 0.0.4